requirements.txt
pandas>=1.5.0
snowflake-snowpark-python>=1.0.0
plotly>=5.0.0
matplotlib>=3.5.0
networkx>=2.8
wordcloud>=1.9.0
streamlit>=1.28.0

